agent-axios-backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
eventlet==0.35.2
flask-cors==4.0.0
sqlalchemy==2.0.23
alembic==1.13.1
faiss-cpu==1.9.0
cohere==4.37
python-dotenv==1.0.0
pydantic==2.5.3
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
gitpython==3.1.40
openai==1.6.1
langsmith
langchain
langchain-openai
langgraph
reportlab==4.0.7
python-socketio[client]==5.10.0
